Refactor, lexical-only.
It turns out that qualifying a class with the class name as a prefix
is not ideal, as it prevents jiggering with the class heierarchy. This
is the inital cut at that work, along with updating to use my
DEFOBJECT class (now no longer in batteries).

Will likely split off the xcore sim-reader into its own world, which
reads a sim and returns a generic container with the appropriate
sexp-styled information. Following that, the generic tta reader will
read it and produce an output.