Rework the sexpr file storage format.
Further, rework the file reading format.

Notes on the current system.

1. Lots of work goes into the sim reader that is related to inferring
up what is already present in a more sophisticated system.

2. The key task next up is to build the edges for each channel. The
new graph should be generated out of the semantic-sexpr-run
information. This graph needs to have all nodes in it and all
channels. Maybe a channel-start for each channel?

Key functions to use to prune are: prune-too-short-paths,
prune-incorrect-sequencing, and prune-bad-starts. See
find-feasible-paths-for-trace as an example.  Note that these should
be passed into the all-paths generation to prune paths on the fly and
not return paths that later need to be pruned.